Refactor files to fit MVC design pattern and add documentation
